Added () and {} support to parsing
Moved a couple of functions to tools.c (retreive_token and add_semicolon)

Added support for () and {} in opcodes that need it. Perhaps the recently added ones need it too, but you can change that easily in the opcodes file.

Further testing is required for this new featuremake clean && clear && make && ./assembleur good_test_prog !